Import of NICOLAW/WWW-Dilbert-1.09 from CPAN.
gitpan-cpan-distribution: WWW-Dilbert
gitpan-cpan-version:      1.09
gitpan-cpan-path:         NICOLAW/WWW-Dilbert-1.09.tar.gz
gitpan-cpan-author:       NICOLAW
gitpan-cpan-maturity:     released

+++ t/01methods.t
@@ -1,4 +1,4 @@
-# $Id: 01methods.t,v 1.8 2004/09/26 18:39:31 nicolaw Exp $
+# $Id: 01methods.t,v 1.9 2004/09/28 06:46:46 nicolaw Exp $
 
 use Test::More qw(no_plan);
 use lib qw(./lib ../lib);
@@ -11,7 +11,8 @@ ok(my $strip_url = $dilbert->get_todays_strip_url(), 'get_todays_strip_url()');
 
 my ($strip1,$width,$height,$file_ext,$strip_id);
 until ($strip1) {
-	ok($strip1 = $dilbert->get_todays_strip_from_website(), 'get_todays_strip_from_website()');
+	eval { $strip1 = $dilbert->get_todays_strip_from_website(); };
+	ok(($@ && $@ =~ /Failed to download strip URL/) || (!$@ && $strip1),'get_todays_strip_from_website()');
 	if ($strip1) {
 		$width = $strip1->width();
 		$height = $strip1->height();
@@ -20,21 +21,25 @@ until ($strip1) {
 	}
 }
 
-ok(my $strip2 = $dilbert->get_strip_from_website($strip_id), "get_strip_from_website(\"$strip_id\")");
+my $strip2;
+eval { $strip2 = $dilbert->get_strip_from_website($strip_id); };
+ok(($@ && $@ =~ /Failed to download strip URL|Unable to convert '\d+' to a valid strip URL/) || (!$@ && $strip2),"get_strip_from_website(\"$strip_id\")");
 
-my $file = "test.$file_ext";
-unlink $file if -e $file;
-open(FH,">$file") || die "Unable to open file handle FH for filename '$file': $!";
-ok(my $strip_blob = $strip2->strip_blob, "Write strip_blob out to $file");
-binmode FH;
-print FH $strip_blob;
-close(FH);
+if ($strip2 && !$@) {
+	my $file = "test.$file_ext";
+	unlink $file if -e $file;
+	open(FH,">$file") || die "Unable to open file handle FH for filename '$file': $!";
+	ok(my $strip_blob = $strip2->strip_blob, "Write strip_blob out to $file");
+	binmode FH;
+	print FH $strip_blob;
+	close(FH);
 
-ok(my $strip3 = $dilbert->get_strip_from_filename($file), "get_strip_from_filename('$file')");
+	ok(my $strip3 = $dilbert->get_strip_from_filename($file), "get_strip_from_filename('$file')");
 
-ok($strip3->width() == $width, 'Check image width');
-ok($strip3->height() == $height, 'Check image height');
-ok($strip3->file_ext() == $file_ext, 'Check image file_ext');
+	ok($strip3->width() == $width, 'Check image width');
+	ok($strip3->height() == $height, 'Check image height');
+	ok($strip3->file_ext() == $file_ext, 'Check image file_ext');
+}
 
 ok(my $strip4 = $dilbert->get_strip_from_filename('dilbert2004081525314.gif'),
 	'get_strip_from_filename("dilbert2004081525314.gif")');
